backend/scripts/requirements.txt
# backend/scripts/requirements.txt
python-pptx
pytesseract
Pillow